Added verbose option for computing coordinates.
The verbose option helps to make the analysis clearer. If not activated,
only data writings, process startings and sensor closings will be shown.
On the other hand, when activated, iteven shows the coordinates and when
it is impossible to compute the homography (minimiz. not reached and not
enough key points).